Move changeAutoTrigger() to protocol instead of extension for UIViewControllers
