+++ Example/LayoutInspectorExample/TableDemoViewController.swift
@@ -9,7 +9,7 @@
 import UIKit
 import LayoutInspector
 
-class TableDemoViewController: UITableViewController {
+class TableDemoViewController: UITableViewController, ChangeAutoTriggerProtocol {
     
     @IBAction private func inspectAction(_ sender: Any) {
         LayoutInspector.shared.showLayout()
@@ -17,20 +17,5 @@ class TableDemoViewController: UITableViewController {
 	
 	@IBAction func changeAutoTrigger(_ sender: Any) {
 		changeAutoTrigger()
-	}
-}
-
-extension UIViewController {
-	
-	func changeAutoTrigger() {
-		let alert = UIAlertController(title: nil, message: "Select auto trigger", preferredStyle: .actionSheet)
-		alert.addAction(UIAlertAction(title: "Screenshot", style: .default, handler: { (_) in
-			LayoutInspector.shared.setAutoTrigger(.screenshot)
-		}))
-		alert.addAction(UIAlertAction(title: "Shake", style: .default, handler: { (_) in
-			LayoutInspector.shared.setAutoTrigger(.shake)
-		}))
-		alert.addAction(UIAlertAction(title: "Cancel", style: .cancel, handler: nil))
-		present(alert, animated: true, completion: nil)
 	}
 }
